Move `Vector` and `Square` to `square.py` as it is needed by both `pieces.py` and `board.py`
